+++ src/scripts/databaseManager/blueprints/culture.ts
@@ -416,6 +416,26 @@ export const cultureBlueprint: I_Blueprint = {
         connectedObjectType: "tech",
         connectedField: "pairedConnectedCultures"
       }
+    },
+    {
+      id: "breakDetails",
+      name: "Secrets/Spoilers/DM notes",
+      type: "break",
+      sizing: 12,
+      isSpoiler: true
+    },
+    {
+      id: "spoilerNotes",
+      name: "Secrets/Spoilers/DM notes",
+      type: "wysiwyg",
+      icon: "fas fa-mask",
+      sizing: 12,
+      isSpoiler: true,
+      tooltip:
+        `This field will not export by default when using the Export document functionality.
+        <br>
+        Instead it needs to be manually included if the user wishes to export it.
+        `
     }
   ]
 }