+++ src/scripts/databaseManager/blueprints/politicalGroups.ts
@@ -699,6 +699,26 @@ export const politicalGroupsBlueprint: I_Blueprint = {
         connectedObjectType: "conditions",
         connectedField: "pairedRacesPoliticalGroups"
       }
+    },
+    {
+      id: "breakDetails",
+      name: "Secrets/Spoilers/DM notes",
+      type: "break",
+      sizing: 12,
+      isSpoiler: true
+    },
+    {
+      id: "spoilerNotes",
+      name: "Secrets/Spoilers/DM notes",
+      type: "wysiwyg",
+      icon: "fas fa-mask",
+      sizing: 12,
+      isSpoiler: true,
+      tooltip:
+        `This field will not export by default when using the Export document functionality.
+        <br>
+        Instead it needs to be manually included if the user wishes to export it.
+        `
     }
 
   ]